workers/requirements.txt
fastapi
uvicorn
httpx
requests
trafilatura
python-dotenv
pydantic
beautifulsoup4

torch
torchvision
diffusers>=0.30.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors
sentence-transformers
Pillow
protobuf
sentencepiece
huggingface_hub

qdrant-client
duckduckgo-search